add delete category test and updated impl

+++ budget-binder-server/src/main/kotlin/de/hsfl/budgetBinder/server/services/implementations/CategoryServiceImpl.kt
@@ -89,8 +89,11 @@ class CategoryServiceImpl : CategoryService {
         categoryEntity.toDto()
     }
 
-    override fun deleteCategory(categoryId: Int): Category = transaction {
+    override fun deleteCategory(categoryId: Int): Category? = transaction {
         val categoryEntity = CategoryEntity[categoryId]
+        if (categoryEntity.ended != null) {
+            return@transaction null
+        }
         categoryEntity.entries.forEach { it.category = CategoryEntity[categoryEntity.user.category!!] }
         val returnValue = categoryEntity.toDto()
         categoryEntity.delete()
